# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "Documentación completa de la API"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/actividades/imagenes": {
      "post": {
        "tags": [
          "Imagenes"
        ],
        "summary": "Guardar imagen",
        "description": "Guarda una imagen en el bucket de actividades asociada a una agencia",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "image"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Archivo de imagen a subir"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Imagen guardada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Imagen guardada exitosamente"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "signed_url": {
                          "type": "string",
                          "format": "uri",
                          "example": "https://example.com/signed-url"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Datos de imagen inválidos"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Imagenes"
        ],
        "summary": "Obtener URL de una imagen",
        "description": "Obtiene la URL firmada de una imagen asociada a una actividad",
        "parameters": [
          {
            "in": "query",
            "name": "actividadId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID de la actividad"
          }
        ],
        "responses": {
          "200": {
            "description": "URL obtenida exitosamente"
          },
          "404": {
            "description": "Imagen no encontrada"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "put": {
        "tags": [
          "Imagenes"
        ],
        "summary": "Actualizar una imagen",
        "description": "Actualiza una imagen asociada a una actividad",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "actividadId",
                  "image"
                ],
                "properties": {
                  "actividadId": {
                    "type": "integer",
                    "description": "ID de la actividad"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Archivo de imagen a subir"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Imagen actualizada exitosamente"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "delete": {
        "tags": [
          "Imagenes"
        ],
        "summary": "Eliminar una imagen",
        "description": "Elimina una imagen asociada a una actividad",
        "parameters": [
          {
            "in": "query",
            "name": "actividadId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID de la actividad"
          }
        ],
        "responses": {
          "200": {
            "description": "Imagen eliminada exitosamente"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
    "/api/usuarios/{id}": {
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Eliminar un usuario",
        "description": "Permite eliminar un usuario específico de una agencia",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del usuario a eliminar"
          },
          {
            "in": "header",
            "name": "x-agencia-id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID de la agencia"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado exitosamente"
          },
          "400": {
            "description": "Datos de entrada inválidos"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
  },
...
}